+++ data/src/jvmMain/kotlin/fobo66/valiutchik/core/model/datasource/FormattingDataSourceIcuImpl.kt
@@ -34,7 +34,7 @@ class FormattingDataSourceIcuImpl(
     /**
      * Format currency rate as a monetary value
      */
-    override fun formatCurrencyValue(value: Double): String = NumberFormatter
+    override fun formatCurrencyValue(value: Float): String = NumberFormatter
         .withLocale(locale)
         .unit(targetCurrency)
         .format(value)
